budgetee-server/requirements.txt
Flask==2.0.3
Flask_Admin==1.6.0
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
psycopg2
PyJWT==2.3.0
pytest==7.1.2
python-dotenv==0.20.0
SQLAlchemy==1.4.32
